requirements.txt
numpy==2.3.5
Pillow==12.0.0
streamlit==1.46.1
tensorflow_cpu==2.18.1
